requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
plotly>=5.15.0
yfinance>=0.2.28
ccxt>=4.0.0
backtrader>=1.9.78.123
ta-lib>=0.4.28
streamlit>=1.25.0
torch>=2.0.0
tushare>=1.2.89
python-dotenv>=1.0.0
google-generativeai>=0.3.0
